+++ src/domain/assignment/assignment-repository.ts
@@ -1,0 +1,6 @@
+import type { Assignment } from "./assignment";
+
+export interface IAssignmentRepository {
+  save(Assignment: Assignment): Promise<void>;
+  findById(id: string): Promise<Assignment | null>;
+}
