--- src/domain/task/task-repository.ts
@@ -1,6 +1,0 @@
-import type { Task } from "./task";
-
-export interface ITaskRepository {
-  save(task: Task): Promise<void>;
-  findById(id: string): Promise<Task | null>;
-}
