+++ src/domain/task/task.ts
@@ -1,48 +1,57 @@
 import { ulid } from "../../libs/ulid";
 
-export class Task {
-  #id: string;
-  #name: string;
-  #content_url: string;
+type AssignmentStatus = "未着手" | "レビュー待ち" | "完了";
 
-  private constructor(id: string, name: string, content_url: string) {
-    if (name.length === 0) {
-      throw new Error("課題名が未入力です。");
-    }
+export class Assignment {
+  #id: string;
+  #studentId: string;
+  #taskId: string;
+  #status: AssignmentStatus;
+
+  private constructor(
+    id: string,
+    studentId: string,
+    taskId: string,
+    status: AssignmentStatus,
+  ) {
     this.#id = id;
-    this.#name = name;
-    this.#content_url = content_url;
+    this.#studentId = studentId;
+    this.#taskId = taskId;
+    this.#status = status;
   }
 
-  public static create(name: string, content_url: string): Task {
+  public static create(studentId: string, taskId: string): Assignment {
     const id = ulid();
-    return new Task(id, name, content_url);
+    const status: AssignmentStatus = "未着手";
+    return new Assignment(id, studentId, taskId, status);
   }
 
   public static reconstruct(
     id: string,
-    name: string,
-    content_url: string,
-  ): Task {
-    return new Task(id, name, content_url);
+    studentId: string,
+    taskId: string,
+    status: AssignmentStatus,
+  ): Assignment {
+    return new Assignment(id, studentId, taskId, status);
   }
 
   public get id(): string {
     return this.#id;
   }
 
-  public get name(): string {
-    return this.#name;
+  public get studentId(): string {
+    return this.#studentId;
   }
 
-  public get content_url(): string {
-    return this.#content_url;
+  public get taskId(): string {
+    return this.#taskId;
   }
 
-  public changeName(newName: string): void {
-    if (newName.length === 0) {
-      throw new Error("課題名が未入力です。");
-    }
-    this.#name = newName;
+  public get status(): AssignmentStatus {
+    return this.#status;
+  }
+
+  public changeStatus(newStatus: AssignmentStatus): void {
+    this.#status = newStatus;
   }
 }
